+++ sdanilov/OOPFirst/MyTriangle.java
@@ -7,7 +7,7 @@ public class MyTriangle {
     private MyPoint v2;
     private MyPoint v3;
 
-    public MyTriangle(int x1, int y1, int x2, int y2, int x3, int y3){
+    public MyTriangle(float x1, float y1, float x2, float y2, float x3, float y3){
         this(new MyPoint(x1,y1), new MyPoint(x2,y2), new MyPoint(x3,y3));
     }
